[Feat] Create new CheckBoxClickable Component
why: make the MarkdownEditor Component cleaner
how: move relevant code to CheckBoxClickable Component file
