+++ src/components/Common/MarkdownEditor/MarkdownEditor.jsx
@@ -8,15 +8,8 @@ import {
 } from "./MarkdownEditorElements";
 import rehypeSanitize from "rehype-sanitize";
 import "./MarkdownEditor.css";
+import CheckBoxClickable from "./CheckBoxClickable";
 
-const compareStrings = (str1, str2) => {
-    for (let i = 0, j = 0; i < str1.length || j < str2.length; i++, j++) {
-        if (i >= str1.length) return false;
-        if (j >= str2.length) return false;
-        if (str1[i] !== str2[j]) return false;
-    }
-    return true;
-};
 const MarkdownEditor = ({ content, label, previewModeOnly, onCopyChanges }) => {
     const [value, setValue] = useState("");
     useEffect(() => {
@@ -30,7 +23,7 @@ const MarkdownEditor = ({ content, label, previewModeOnly, onCopyChanges }) => {
                 style={{ whiteSpace: "normal", backgroundColor: "#000" }}
                 components={{
                     input: (props) => {
-                        return <input {...props} disabled={true} />;
+                        return <CheckBoxClickable disabled={true} {...props} />;
                     },
                 }}
             />
@@ -41,18 +34,6 @@ const MarkdownEditor = ({ content, label, previewModeOnly, onCopyChanges }) => {
         onCopyChanges(label, value);
     };
 
-    const handleCheckBoxChange = (e) => {
-        const textOfCheckBox = e.target.parentNode.textContent;
-        const valueListOfLines = value.split("\n");
-        const findCheckedBoxLineIndex = valueListOfLines.findIndex((item) =>
-            compareStrings(item?.replace(/- \[ \]|- \[[^]]+/, ""), textOfCheckBox.split("\n")[0]),
-        );
-        valueListOfLines[findCheckedBoxLineIndex] = valueListOfLines[findCheckedBoxLineIndex].replace(
-            /- \[ \]|- \[[^]]+/,
-            (match) => (match === "- [ ]" ? "- [X]" : "- [ ]"),
-        );
-        handleChange(valueListOfLines.join("\n"));
-    };
     return (
         <MarkdownContainer>
             <MarkdownLabel>{label}</MarkdownLabel>
@@ -67,20 +48,11 @@ const MarkdownEditor = ({ content, label, previewModeOnly, onCopyChanges }) => {
                     components={{
                         input: (props) => {
                             return (
-                                <input
+                                <CheckBoxClickable
                                     {...props}
                                     disabled={false}
-                                    onChange={(e) => {
-                                        if (props.type === "checkbox") {
-                                            const isChecked = e.target.hasAttribute("checked");
-                                            handleCheckBoxChange(e);
-                                            if (isChecked) {
-                                                e.target.removeAttribute("checked");
-                                            } else {
-                                                e.target.setAttribute("checked", "checked");
-                                            }
-                                        }
-                                    }}
+                                    value={value}
+                                    onChangeValue={handleChange}
                                 />
                             );
                         },
